[refactor] #25 domain/common 디렉토리 생성, Track 공통 Enum 적용
