+++ src/main/java/org/farmsystem/homepage/domain/apply/dto/QuestionDTO.java
@@ -1,12 +1,12 @@
 package org.farmsystem.homepage.domain.apply.dto;
 
 import lombok.Builder;
-import org.farmsystem.homepage.domain.apply.entity.ApplyTrack;
 import org.farmsystem.homepage.domain.apply.entity.QuestionType;
+import org.farmsystem.homepage.domain.common.entity.Track;
 
 import java.util.List;
 
 @Builder
-public record QuestionDTO(Long questionId, ApplyTrack track, Boolean isRequired, String content, Integer maxLength,
+public record QuestionDTO(Long questionId, Track track, Boolean isRequired, String content, Integer maxLength,
                           QuestionType type, Boolean isDuplicated, Integer priority, List<ChoiceDTO> choices) {
 }
