--- src/main/java/org/farmsystem/homepage/domain/apply/entity/ApplyTrack.java
@@ -1,5 +1,0 @@
-package org.farmsystem.homepage.domain.apply.entity;
-
-public enum ApplyTrack {
-    UNION, GAMING_VIDEO, SECURITY_WEB, AI, IOT_ROBOTICS, BIGDATA
-}
