+++ src/main/java/org/farmsystem/homepage/domain/common/entity/Track.java
@@ -1,4 +1,4 @@
-package org.farmsystem.homepage.domain.user.entity;
+package org.farmsystem.homepage.domain.common.entity;
 
 public enum Track {
     UNION, GAMING_VIDEO, SECURITY_WEB, AI, IOT_ROBOTICS, BIGDATA
